kube-bot/requirements.txt
kubernetes>=26.1.0
slack-sdk>=3.26.0
PyYAML>=6.0
